move warn into util
